refactor: move visit_*_bis shared by 1st and 3rd passes into common module
